feat(backend,synthese) : divide synthese routes in multiple files
