+++ backend/geonature/core/gn_synthese/blueprints/__init__.py
@@ -1,0 +1,6 @@
+from geonature.core.gn_synthese.blueprints.reports import reports_blueprint
+from geonature.core.gn_synthese.blueprints.synthese import synthese_routes
+from geonature.core.gn_synthese.blueprints.statistics import statistics_routes
+from geonature.core.gn_synthese.blueprints.taxon_info import taxon_info_routes
+from geonature.core.gn_synthese.blueprints.others import other_routes
+from geonature.core.gn_synthese.blueprints.exports import export_routes
